+++ src/Config.ts
@@ -660,6 +660,7 @@ export class Config implements ConfigParams, ParserConfig {
     return ret as ConfigParams
   }
 
+
   public mergeConfig(init: Partial<ConfigParams>): Config {
     const mergedConfig: ConfigParams = Object.assign({}, this.getConfig(), init)