+++ src/CrudOperations.ts
@@ -29,7 +29,6 @@ import {
   NoOperationToUndoError,
   NoRelativeAddressesAllowedError,
   NoSheetWithIdError,
-  NoSheetWithNameError,
   NothingToPasteError,
   SheetNameAlreadyTakenError,
   SheetSizeLimitExceededError,
@@ -214,14 +213,13 @@ export class CrudOperations {
     return addedSheetName
   }
 
-  public removeSheet(sheetName: string): void {
-    this.ensureSheetExists(sheetName)
+  public removeSheet(sheetId: number): void {
+    this.ensureScopeIdIsValid(sheetId)
     this.undoRedo.clearRedoStack()
     this.clipboardOperations.abortCut()
-    const sheetId = this.sheetMapping.fetch(sheetName)
     const originalName = this.sheetMapping.fetchDisplayName(sheetId)
     const oldSheetContent = this.operations.getSheetClipboardCells(sheetId)
-    const {version, scopedNamedExpressions} = this.operations.removeSheet(sheetName)
+    const {version, scopedNamedExpressions} = this.operations.removeSheet(sheetId)
     this.undoRedo.saveOperation(new RemoveSheetUndoEntry(originalName, sheetId, oldSheetContent, scopedNamedExpressions, version))
   }
 
@@ -235,11 +233,10 @@ export class CrudOperations {
     return oldName
   }
 
-  public clearSheet(sheetName: string): void {
-    this.ensureSheetExists(sheetName)
+  public clearSheet(sheetId: number): void {
+    this.ensureScopeIdIsValid(sheetId)
     this.undoRedo.clearRedoStack()
     this.clipboardOperations.abortCut()
-    const sheetId = this.sheetMapping.fetch(sheetName)
     const oldSheetContent = this.operations.getSheetClipboardCells(sheetId)
     this.operations.clearSheet(sheetId)
     this.undoRedo.saveOperation(new ClearSheetUndoEntry(sheetId, oldSheetContent))
@@ -276,9 +273,8 @@ export class CrudOperations {
     this.undoRedo.saveOperation(new SetCellContentsUndoEntry(modifiedCellContents))
   }
 
-  public setSheetContent(sheetName: string, values: RawCellContent[][]): void {
-    this.ensureSheetExists(sheetName)
-    const sheetId = this.sheetMapping.fetch(sheetName)
+  public setSheetContent(sheetId: number, values: RawCellContent[][]): void {
+    this.ensureScopeIdIsValid(sheetId)
     this.ensureItIsPossibleToChangeSheetContents(sheetId, values)
 
     validateAsSheet(values)
@@ -396,8 +392,8 @@ export class CrudOperations {
     this.undoRedo.redo()
   }
 
-  public addNamedExpression(expressionName: string, expression: RawCellContent, sheetScope?: string, options?: NamedExpressionOptions) {
-    const sheetId = this.scopeId(sheetScope)
+  public addNamedExpression(expressionName: string, expression: RawCellContent, sheetId?: number, options?: NamedExpressionOptions) {
+    this.ensureScopeIdIsValid(sheetId)
     this.ensureNamedExpressionNameIsValid(expressionName, sheetId)
     this.operations.addNamedExpression(expressionName, expression, sheetId, options)
     this.undoRedo.clearRedoStack()
@@ -405,16 +401,16 @@ export class CrudOperations {
     this.undoRedo.saveOperation(new AddNamedExpressionUndoEntry(expressionName, expression, sheetId, options))
   }
 
-  public changeNamedExpressionExpression(expressionName: string, sheetScope: string | undefined, newExpression: RawCellContent, options?: NamedExpressionOptions) {
-    const sheetId = this.scopeId(sheetScope)
+  public changeNamedExpressionExpression(expressionName: string, sheetId: number | undefined, newExpression: RawCellContent, options?: NamedExpressionOptions) {
+    this.ensureScopeIdIsValid(sheetId)
     const [oldNamedExpression, content] = this.operations.changeNamedExpressionExpression(expressionName, newExpression, sheetId, options)
     this.undoRedo.clearRedoStack()
     this.clipboardOperations.abortCut()
     this.undoRedo.saveOperation(new ChangeNamedExpressionUndoEntry(oldNamedExpression, newExpression, content, sheetId, options))
   }
 
-  public removeNamedExpression(expressionName: string, sheetScope: string | undefined): InternalNamedExpression {
-    const sheetId = this.scopeId(sheetScope)
+  public removeNamedExpression(expressionName: string, sheetId?: number): InternalNamedExpression {
+    this.ensureScopeIdIsValid(sheetId)
     const [namedExpression, content] = this.operations.removeNamedExpression(expressionName, sheetId)
     this.undoRedo.clearRedoStack()
     this.clipboardOperations.abortCut()
@@ -423,23 +419,23 @@ export class CrudOperations {
     return namedExpression
   }
 
-  public ensureItIsPossibleToAddNamedExpression(expressionName: string, expression: RawCellContent, sheetScope?: string): void {
-    const scopeId = this.scopeId(sheetScope)
-    this.ensureNamedExpressionNameIsValid(expressionName, scopeId)
+  public ensureItIsPossibleToAddNamedExpression(expressionName: string, expression: RawCellContent, sheetId?: number): void {
+    this.ensureScopeIdIsValid(sheetId)
+    this.ensureNamedExpressionNameIsValid(expressionName, sheetId)
     this.ensureNamedExpressionIsValid(expression)
   }
 
-  public ensureItIsPossibleToChangeNamedExpression(expressionName: string, expression: RawCellContent, sheetScope?: string): void {
-    const scopeId = this.scopeId(sheetScope)
-    if (this.namedExpressions.namedExpressionForScope(expressionName, scopeId) === undefined) {
+  public ensureItIsPossibleToChangeNamedExpression(expressionName: string, expression: RawCellContent, sheetId?: number): void {
+    this.ensureScopeIdIsValid(sheetId)
+    if (this.namedExpressions.namedExpressionForScope(expressionName, sheetId) === undefined) {
       throw new NamedExpressionDoesNotExistError(expressionName)
     }
     this.ensureNamedExpressionIsValid(expression)
   }
 
-  public isItPossibleToRemoveNamedExpression(expressionName: string, sheetScope?: string): void {
-    const scopeId = this.scopeId(sheetScope)
-    if (this.namedExpressions.namedExpressionForScope(expressionName, scopeId) === undefined) {
+  public isItPossibleToRemoveNamedExpression(expressionName: string, sheetId?: number): void {
+    this.ensureScopeIdIsValid(sheetId)
+    if (this.namedExpressions.namedExpressionForScope(expressionName, sheetId) === undefined) {
       throw new NamedExpressionDoesNotExistError(expressionName)
     }
   }
@@ -648,18 +644,10 @@ export class CrudOperations {
     return this.operations.getAndClearContentChanges()
   }
 
-  public ensureSheetExists(sheetName: string): void {
-    if (!this.sheetMapping.hasSheetWithName(sheetName)) {
-      throw new NoSheetWithNameError(sheetName)
+  public ensureScopeIdIsValid(scopeId?: number): void {
+    if(scopeId !== undefined && !this.sheetMapping.hasSheetWithId(scopeId)) {
+      throw new NoSheetWithIdError(scopeId)
     }
-  }
-
-  public scopeId(sheetName?: string): Maybe<number> {
-    if (sheetName !== undefined) {
-      this.ensureSheetExists(sheetName)
-      return this.sheetMapping.fetch(sheetName)
-    }
-    return undefined
   }
 
   private get sheetMapping(): SheetMapping {