+++ test/rebuild.spec.ts
@@ -15,7 +15,7 @@ describe('Rebuilding engine', () => {
       'Sheet1': [['=FALSE']],
       'Sheet2': [['=FALSE']]
     })
-    engine.addNamedExpression('FALSE', '=FALSE()', 'Sheet1')
+    engine.addNamedExpression('FALSE', '=FALSE()', 0)
     engine.rebuildAndRecalculate()
     expect(engine.getCellValue(adr('A1', 0))).toEqual(false)
     expect(engine.getCellValue(adr('A1', 1))).toEqualError(detailedError(ErrorType.NAME, ErrorMessage.NamedExpressionName('FALSE')))